backend/requirements.txt
fastapi
uvicorn[standard]
openai
anthropic
python-dotenv
pydantic
pypdf
httpx
python-multipart
beautifulsoup4

langchain
langchain-openai
langchain-anthropic
